requirements.txt
asgiref==3.2.3
attrs==19.3.0
boto==2.49.0
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-countries==5.5
django-dotenv==1.4.2
django-formtools==2.2
django-lockdown==3.0.0
django-ses==0.8.14
djangorestframework==3.11.0
djangorestframework-gis==0.15
drf-yasg==1.17.1
future==0.18.2
gunicorn==20.0.4
honcho==1.0.1
idna==2.8
importlib-metadata==1.4.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.1.0
oauthlib==3.1.0
packaging==20.0
psycopg2==2.8.4
pyparsing==2.4.6
pyrsistent==0.15.7
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
zipp==1.0.0
